1.1.0: Make `connector` return enumerable properties.
Updated the documentation to reflect our new-found preference for enumerable properties as state inspectors, as opposed to methods on the Api object. Enumerable properties better reflect the read-only nature of state inspectors and they play ball very well with ES6 spread / destructuring syntax.
Update examples to use new style.

This is a non-breaking change, as reflected in the major version not changing, but we did deprecate the old behavior of `connector`, so stop relying on it to prepare for v2.x, in which this behavior will be removed.